optimized iphone 4s

+++ GameTimeKill/ViewController.swift
@@ -306,8 +306,8 @@ extension String{
 }
 
 extension Date {
-    var millisecondsSince1970:Int {
-        return Int((self.timeIntervalSince1970 * 1000.0).rounded())
+    var millisecondsSince1970:Int64 {
+        return Int64((self.timeIntervalSince1970 * 1000.0).rounded())
     }
     
     init(milliseconds:Int) {
